Split DotToDotEditorClient into hook + focused components
The editor page mixed state, canvas-tracing handlers, and four unrelated
JSX sections in one 287-line file. Extracted the state/handlers into
useDotToDotEditor, the pure code-gen helper into buildPictureCode, and the
JSX sections into components/ (PointCanvas, EditorControls, MetadataForm,
ExportPanel), matching the app/games/<game>/components/ shape used
elsewhere in the repo. No behavior change.
