+++ gamesformykids/app/games/dot-to-dot/editor/editorConstants.ts
@@ -1,0 +1,7 @@
+export const VIEW_SIZE = 300;
+export const DISPLAY_WIDTH = 500;
+export const MAX_WORKING_DIMENSION = 500;
+export const MIN_DOTS = 8;
+export const MAX_DOTS = 60;
+export const DEFAULT_DOTS = 20;
+export const SIMPLIFY_EPSILON_RATIO = 0.03;
